[node-api-maps_distance] distance の処理を別クラスに定義
