+++ src/node/index.ts
@@ -3,6 +3,7 @@ import swaggerUi from 'swagger-ui-express'
 import yamljs from 'yamljs'
 
 import { maps, mapsLatLon } from '../ts/maps';
+import { apiMapsDistance } from './apiMapsDistance';
 
 const app:express.Express = express();
 
@@ -123,91 +124,8 @@ router.get('/api/maps/jgd2tkyg',
 	}
 );
 
-interface mapsData_Distance {
-	status: boolean
-	, distance: number
-};
-/**
- * ２地点間の距離
- * @param type T:球面三角法,H:ヒュベニ,S:測地線航海算法
- * @param req リクエスト
- * @returns 結果
- */
-const apiMapsDistance = (type: string, req:express.Request): mapsData_Distance => {
-	let data: mapsData_Distance = {
-		status: false
-		, distance: 0
-	};
-
-	const oMaps: maps = new maps();
-	
-	if (type === "T" || type === "H" || type === "S") {
-		if (req.query.lat1 && req.query.lon1 && req.query.lat2 && req.query.lon2) {
-			const lat1: number = +req.query.lat1;
-			const lon1: number = +req.query.lon1;
-			const lat2: number = +req.query.lat2;
-			const lon2: number = +req.query.lon2;
-			if (!Number.isNaN(lat1) && !Number.isNaN(lon1) && !Number.isNaN(lat2) && !Number.isNaN(lon2)) {
-				if (type === "T") {
-					const distance: number = oMaps.distanceT(lat1, lon1, lat2, lon2);
-					data.status = true;
-					data.distance = distance;
-				}
-				else if (type === "H") {
-					const distance: number = oMaps.distanceH(lat1, lon1, lat2, lon2);
-					data.status = true;
-					data.distance = distance;
-				}
-				else if (type === "S") {
-					const distance: number = oMaps.distanceS(lat1, lon1, lat2, lon2);
-					data.status = true;
-					data.distance = distance;
-				}
-			}
-		}
-	}
-	return data;
-}
-
-router.get('/api/maps/distancet',
-	(req:express.Request, res:express.Response) => {
-		let data: mapsData_Distance = {
-			status: false
-			, distance: 0
-		};
-		data = apiMapsDistance("T", req);
-		
-		res.json(data);
-		res.end();
-	}
-);
-
-router.get('/api/maps/distanceh',
-	(req:express.Request, res:express.Response) => {
-		let data: mapsData_Distance = {
-			status: false
-			, distance: 0
-		};
-		data = apiMapsDistance("H", req);
-		
-		res.json(data);
-		res.end();
-	}
-);
-
-router.get('/api/maps/distances',
-	(req:express.Request, res:express.Response) => {
-		let data: mapsData_Distance = {
-			status: false
-			, distance: 0
-		};
-		data = apiMapsDistance("S", req);
-		
-		res.json(data);
-		res.end();
-	}
-);
-
+const oApiMapsDistance = new apiMapsDistance();
+oApiMapsDistance.regist(router);
 
 app.use(router);
 app.use('/', express.static('public'));
